:ok_hand: Move setting validator to model file
