+++ src/NzbDrone.Core/Download/Clients/RQBit/RQbitSettings.cs
@@ -33,4 +33,15 @@ namespace NzbDrone.Core.Download.Clients.RQBit
             return new NzbDroneValidationResult(Validator.Validate(this));
         }
     }
+
+    public class RQbitSettingsValidator : AbstractValidator<RQbitSettings>
+    {
+        public RQbitSettingsValidator()
+        {
+            RuleFor(c => c.Host).ValidHost();
+            RuleFor(c => c.Port).InclusiveBetween(1, 65535);
+
+            RuleFor(c => c.UrlBase).ValidUrlBase();
+        }
+    }
 }
